clean: UserInputStateを移動
・stateディレクトリにuser_input_state.dartファイルを作り、そこへUserInputクラスを移動させた。
・以前に作成したSignUpStateは不要でコメントアウトされていたが、それを削除した。
